requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
plotly>=5.13.0
streamlit>=1.22.0
tqdm>=4.65.0
pyyaml>=6.0
pytest>=7.3.1
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 